common directory with getline function realization
